requirements.txt
pypdf
python-docx
requests